backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
google-api-python-client==2.130.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gTTS==2.5.1
cryptography==42.0.7
python-jose[cryptography]==3.3.0
pydantic==2.7.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
